Do not add charset to application/x-www-form-urlencoded header.
Adding a charset to this Content-Type is illegal according to the definition, and fails with Zope master.
See https://github.com/plone/buildout.coredev/pull/844#issuecomment-1461620409

In our case, when you click to add a recurrence, you get a dialog box saying 'error' because the call to `@@json_recurrence` failed.